Add MainSerializer class
This class was made for ease of use since it registers all custom adapters in a single place and provides a handy Gson instance.
Current theory is to have a Companion Object that holds an MainSerializer on the Mod's entry class so that serialization can safely be done wherever it's needed once we implement io.

If there is a better way to do that please correct me since I'm not yet to speed in idiomatic kotlin.